move functions to remove.js
